pyproject.toml
[project]
name = "abbonamenti"
version = "0.3.0.7"
description = "Secure subscription archiving system for municipal police office"
readme = "README.md"
requires-python = ">=3.8,<3.9"
dependencies = [
    "cryptography==3.4.8",
    "matplotlib==3.5.3",
    "openpyxl==3.0.10",
    "pyinstaller==5.13.2",
    "python-telegram-bot==13.15",
    "pyqt5==5.15.2",
    "reportlab==3.6.12",
]

[project.scripts]
abbonamenti = "abbonamenti.main:main"
abbonamenti-bot = "abbonamenti.bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
